requirements.txt
requests
pycrypto
Flask